requirements.txt
google-adk[extensions]>=1.26.0
google-genai>=1.0.0
python-dotenv>=1.0.1
